vit_requirements.txt
absl-py==1.4.0
addict==2.4.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==3.1.0
chumpy==0.70
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
einops==0.6.1
fonttools==4.39.4
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.54.0
hdf5storage==0.1.19
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
json-tricks==3.16.1
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
#-e git+https://github.com/open-mmlab/mmcv.git@96c4b70ccbb9cbf452ff79695a35187f919b0c9d#egg=mmcv_full
munkres==1.1.4
ninja==1.11.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==3.20.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
scikit-image==0.21.0
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
timm==0.4.9
tomli==2.0.1
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.4
xtcocotools==1.13
yapf==0.33.0
zipp==3.15.0
